BackEND/requirements.txt
anyio==3.6.2
boto3==1.26.22
botocore==1.29.22
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
pymongo-auth-aws==1.1.0
pymongocrypt==1.4.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
